+++ packages/admin-stories/src/admin/table/TableEditDialogHooks.tsx
@@ -1,7 +1,7 @@
 import {
     Field,
     FinalForm,
-    FinalFormTextField,
+    FinalFormInput,
     MainContent,
     Selected,
     Table,
@@ -38,7 +38,7 @@ function EditForm(props: IEditFormProps) {
                 alert(JSON.stringify(values));
             }}
         >
-            <Field name="foo" component={FinalFormTextField} type="text" label="Name" />
+            <Field name="foo" component={FinalFormInput} type="text" label="Name" fullWidth />
         </FinalForm>
     );
 }