+++ packages/admin/src/common/buttons/save/SaveButton.tsx
@@ -25,8 +25,8 @@ export const SaveButton = ({
     savingItem = <FormattedMessage id={"comet.saveButton.savingItem.title"} defaultMessage={"Saving"} />,
     successItem = <FormattedMessage id={"comet.saveButton.successItem.title"} defaultMessage={"Success Saved"} />,
     errorItem = <FormattedMessage id={"comet.saveButton.errorItem.title"} defaultMessage={"Save Error"} />,
-    variant,
-    color,
+    variant = "contained",
+    color = "primary",
     classes: passedClasses,
     ...restProps
 }: PropsWithChildren<SaveButtonProps> & StyledComponentProps<CometAdminSaveButtonClassKeys>) => {