+++ CHANGELOG.md
@@ -30,6 +30,9 @@ All notable changes to this project will be documented in this file. This projec
 -   add `FinalFormSaveCancelButtonsLegacy` as drop in replacement for removed Cancel and Save Button in `FinalForm`.
 -   add PrettyBytes component for formatting file sizes and other byte values
 -   Add `validateWarning` validator to `Field` and `FinalForm`.
+-   Add `open` and `onOpenChange` props to `AppHeaderDropdown` that allow replacing the internal open state with an external state
+-   add `getTargetUrl()` to `StackSwitchApi`
+-   add `StackLink` component for navigating within a `Stack` via hyperlinks
 
 ### Incompatible Changes